requirements.txt
Django==2.2.4
django-hosts==3.0
djangorestframework==3.10.2
pytz==2019.2
sqlparse==0.3.0
